split Extensions class
