+++ cs/TagsCloudVisualization/TagsCloudVisualization/IRectangleLayouter.cs
@@ -1,7 +1,9 @@
-﻿namespace TagsCloudVisualization
+﻿using System.Drawing;
+
+namespace TagsCloudVisualization
 {
     public interface IRectangleLayouter
     {
-        
+        public Rectangle PutNextRectangle(Size rectangleSize);
     }
 }
